PYTHONPATH has top level of project
PYTHONPATH imports the top level of the project instead of src and test separately
This is so we don't have to worry about name collisions between src and test

Imports of these will now look like "import src.<module>" and "import test.<module>"
as opposed to just "import <module"
