+++ install.sh
@@ -1,12 +1,11 @@
 #!/bin/bash
+# Must run this script using dot (i.e. . ./install.sh instead of just ./install.sh)
+# This is so that the source on the last line is applied to the parent's shell
+# environment instead of just this script's subshell environment
+
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-printf "PYTHONPATH=\${PYTHONPATH}:$DIR/src\n" >> ~/.profile
-printf "PYTHONPATH=\${PYTHONPATH}:$DIR/test\n" >> ~/.profile
-printf "export PYTHONPATH\n" >> ~/.profile
-source ~/.profile
-
-printf "PYTHONPATH=\${PYTHONPATH}:$DIR/src\n" >> ~/.bash_profile
-printf "PYTHONPATH=\${PYTHONPATH}:$DIR/test\n" >> ~/.bash_profile
+printf "PYTHONPATH=\${PYTHONPATH}:$DIR\n" >> ~/.bash_profile
 printf "export PYTHONPATH\n" >> ~/.bash_profile
+
 source ~/.bash_profile
